(+) refactor unit test
