+++ src/__tests__/Either.test.ts
@@ -3,16 +3,7 @@ import * as w from "winston";
 import { pipe } from "fp-ts/lib/function";
 import * as E from "fp-ts/Either";
 import * as EL from "../Either";
-
-class DummyTransport extends Transport {
-  log = jest.fn();
-}
-const dummyTransport = new DummyTransport();
-
-w.configure({
-  level: "debug",
-  transports: [dummyTransport]
-});
+import { useDummyTransport } from "../../__mocks__/transports";
 
 const dummyMessage = "dummy-message";
 const dummyItem = "dummy-item";
@@ -23,11 +14,13 @@ describe("index", () => {
   });
 
   it("do not log with left", () => {
+    const dummyTransport = useDummyTransport();
     pipe(dummyItem, E.left, EL.log("info", dummyMessage));
     expect(dummyTransport.log).toBeCalledTimes(0);
   });
 
   it("log with string message", () => {
+    const dummyTransport = useDummyTransport();
     const result = pipe(dummyItem, E.right, EL.log("info", dummyMessage));
     expect(result).toEqual(expect.objectContaining({ right: dummyItem }));
     expect(dummyTransport.log).toBeCalledWith(
@@ -37,6 +30,7 @@ describe("index", () => {
   });
 
   it("log with function message", () => {
+    const dummyTransport = useDummyTransport();
     const result = pipe(
       dummyItem,
       E.right,
@@ -53,11 +47,13 @@ describe("index", () => {
   });
 
   it("do not logLeft with right", () => {
+    const dummyTransport = useDummyTransport();
     pipe(dummyItem, E.right, EL.logLeft("info", dummyMessage));
     expect(dummyTransport.log).toBeCalledTimes(0);
   });
 
   it("logLeft with string message", () => {
+    const dummyTransport = useDummyTransport();
     const result = pipe(dummyItem, E.left, EL.logLeft("info", dummyMessage));
     expect(result).toEqual(expect.objectContaining({ left: dummyItem }));
     expect(dummyTransport.log).toBeCalledWith(
